Rename files to match class name and add back Partition.hpp for backward compatibility
